+++ NEWS.md
@@ -1,3 +1,5 @@
+# bbk (development version)
+
 # bbk 0.7.0
 
 * Switch to checkmate for input validation.