+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Add missing data in `bbk_series()`.
 * Support for exchange rates from the Bank of Canada (BoC).
+* Support for Bank of Canada (BoC) data.
 
 # bbk 0.7.0